+++ tests/llm/prompt/test_prompt.py
@@ -33,7 +33,7 @@ class TestPrompt(TestCase):
             .divider()
             .array(items=['item1', 'item2'])
             .array_random_order(items=['item1', 'item2'])
-            .file(file_path=pathlib.Path(__file__).parent / 'test_prompt_information.md')
+            .file(file_path=pathlib.Path(__file__).parent / 'test_document.md')
             .heading(heading='Heading Followed by a line break')
             .line_break()
             .list(items=['item1 after a line break', 'item2'])